remove type based solvers
