+++ src/solvers/cg_solver_stages.hpp
@@ -24,8 +24,6 @@
 #include "interface/meshblock_data.hpp"
 #include "interface/state_descriptor.hpp"
 #include "kokkos_abstraction.hpp"
-#include "solvers/cg_solver.hpp"
-#include "solvers/mg_solver.hpp"
 #include "solvers/mg_solver_stages.hpp"
 #include "solvers/solver_base.hpp"
 #include "solvers/solver_utils_stages.hpp"
@@ -35,6 +33,26 @@
 namespace parthenon {
 
 namespace solvers {
+
+struct CGParams {
+  MGParams mg_params;
+  int max_iters = 1000;
+  std::shared_ptr<Real> residual_tolerance = std::make_shared<Real>(1.e-12);
+  bool precondition = true;
+  bool print_per_step = false;
+  bool relative_residual = false;
+  CGParams() = default;
+  CGParams(ParameterInput *pin, const std::string &input_block) {
+    max_iters = pin->GetOrAddInteger(input_block, "max_iterations", max_iters);
+    *residual_tolerance =
+        pin->GetOrAddReal(input_block, "residual_tolerance", *residual_tolerance);
+    precondition = pin->GetOrAddBoolean(input_block, "precondition", precondition);
+    print_per_step = pin->GetOrAddBoolean(input_block, "print_per_step", print_per_step);
+    mg_params = MGParams(pin, input_block);
+    relative_residual =
+        pin->GetOrAddBoolean(input_block, "relative_residual", relative_residual);
+  }
+};
 
 // The equations class must include a template method
 //
