+++ src/solvers/mg_solver_stages.hpp
@@ -36,6 +36,30 @@ namespace parthenon {
 
 namespace solvers {
 
+struct MGParams {
+  int max_iters = 1000;
+  Real residual_tolerance = 1.e-12;
+  bool do_FAS = true;
+  std::string smoother = "SRJ2";
+  bool two_by_two_diagonal = false;
+  int max_coarsenings = std::numeric_limits<int>::max();
+  std::string prolongation = "OldLinear";
+
+  MGParams() = default;
+  MGParams(ParameterInput *pin, const std::string &input_block) {
+    max_iters = pin->GetOrAddInteger(input_block, "max_iterations", max_iters);
+    residual_tolerance =
+        pin->GetOrAddReal(input_block, "residual_tolerance", residual_tolerance);
+    do_FAS = pin->GetOrAddBoolean(input_block, "do_FAS", do_FAS);
+    smoother = pin->GetOrAddString(input_block, "smoother", smoother);
+    prolongation = pin->GetOrAddString(input_block, "prolongation", prolongation);
+    two_by_two_diagonal =
+        pin->GetOrAddBoolean(input_block, "two_by_two_diagonal", two_by_two_diagonal);
+    max_coarsenings =
+        pin->GetOrAddInteger(input_block, "max_coarsenings", max_coarsenings);
+  }
+};
+
 // The equations_t class must include a template method
 //
 //   template <class x_t, class y_t, class TL_t>
